+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # master
 
 * Add `ReportCursorPosition` sequence
+* Add ANSI escape sequence parser (feature `parser`, disabled by default)
+  * Add `Parser` engine
+  * Add `KeyCode`, `KeyModifiers`, `Mouse`, `MouseButton`, `Sequence` types
 
 # Version 0.1.5